+++ projects/form-elements/src/exports.ts
@@ -2,29 +2,16 @@
  * Public API Surface
  */
 export { ModuleOptions, OPTIONS } from './lib/options'
-export {
-    FValidationConfig,
-    FExplanationConfig,
-    FRowLayouts,
-    FColumnAlignments,
-    FSelectOption,
-    FSelectPlaceholder,
-    FRadiobuttonOption,
-    FButtonType,
-} from './lib/typings'
 export { FormElementsModule } from './lib/form-elements.module'
-export { FButtonComponent } from './lib/f-button/f-button.component'
-export { FCheckboxComponent } from './lib/f-checkbox/f-checkbox.component'
-export { FColumnComponent } from './lib/f-column/f-column.component'
-export { FDateComponent } from './lib/f-date/f-date.component'
-export { FEmailComponent } from './lib/f-email/f-email.component'
-export { FErrorComponent } from './lib/f-error/f-error.component'
-export { FExplanationComponent } from './lib/f-explanation/f-explanation.component'
-export { FIconComponent } from './lib/f-icon/f-icon.component'
-export { FInputComponent } from './lib/f-input/f-input.component'
-export { FPasswordComponent } from './lib/f-password/f-password.component'
-export { FPhoneComponent } from './lib/f-phone/f-phone.component'
-export { FRadiobuttonComponent } from './lib/f-radiobutton/f-radiobutton.component'
-export { FRowComponent } from './lib/f-row/f-row.component'
-export { FSelectComponent } from './lib/f-select/f-select.component'
-export { FTextAreaComponent } from './lib/f-text-area/f-text-area.component'
+export * from './lib/f-button/f-button.component'
+export * from './lib/f-checkbox/f-checkbox.component'
+export * from './lib/f-column/f-column.component'
+export * from './lib/f-date/f-date.component'
+export * from './lib/f-email/f-email.component'
+export * from './lib/f-input/f-input.component'
+export * from './lib/f-password/f-password.component'
+export * from './lib/f-phone/f-phone.component'
+export * from './lib/f-radiobutton/f-radiobutton.component'
+export * from './lib/f-row/f-row.component'
+export * from './lib/f-select/f-select.component'
+export * from './lib/f-text-area/f-text-area.component'