+++ projects/form-elements/src/lib/f-button/f-button.component.ts
@@ -1,7 +1,10 @@
 import { Component, Input, Output, EventEmitter, ChangeDetectionStrategy } from '@angular/core'
 import { modifiers } from '@pascaliske/html-helpers'
-import { FButtonType } from '../typings'
+import { IconDefinition } from '@fortawesome/fontawesome-svg-core'
 
+/**
+ * F-Button
+ */
 @Component({
     selector: 'cmp-f-button',
     templateUrl: './f-button.component.html',
@@ -11,7 +14,7 @@ export class FButtonComponent {
     public static readonly cmpName: string = 'FButtonComponent'
 
     @Input()
-    public type: FButtonType = 'button'
+    public type: 'button' | 'submit' = 'button'
 
     @Input()
     public id: string
@@ -20,18 +23,16 @@ export class FButtonComponent {
     public label: string
 
     @Input()
-    public icon: string
+    public icon: IconDefinition
 
     @Input()
-    public theme = ''
+    public theme: string = 'primary'
 
     @Input()
-    public disabled = false
+    public disabled: boolean = false
 
     @Output()
     public clicked: EventEmitter<Event> = new EventEmitter()
-
-    public constructor() {}
 
     public get classes(): string {
         return modifiers('cmp-f-button', this.theme, {