+++ projects/form-elements/src/lib/f-date/f-date.stories.ts
@@ -1,0 +1,17 @@
+import { storiesOf } from '@storybook/angular'
+import { text } from '@storybook/addon-knobs'
+
+import { FDateComponent } from './f-date.component'
+import FDateReadme from './f-date.readme.md'
+import '../main.scss'
+
+storiesOf('Advanced|F-Date', module)
+    .addParameters({ notes: FDateReadme })
+    .add('Basic', () => ({
+        component: FDateComponent,
+        props: {
+            name: text('name', 'date'),
+            label: text('label', 'Date'),
+            autocomplete: 'off',
+        },
+    }))