+++ CHANGELOG.md
@@ -1,3 +1,56 @@
+# [2.0.0](https://github.com/pascaliske/form-elements/compare/v1.2.5...v2.0.0) (2019-04-03)
+
+
+### Bug Fixes
+
+* **demo:** update demo app ([bd9bd73](https://github.com/pascaliske/form-elements/commit/bd9bd73))
+* **exports:** fix exports of components ([66e34a6](https://github.com/pascaliske/form-elements/commit/66e34a6))
+* **f-button,f-checkbox,f-date:** fix story hierachy separator ([057507b](https://github.com/pascaliske/form-elements/commit/057507b))
+* **f-date:** fix condition for disabling date picker ([496e045](https://github.com/pascaliske/form-elements/commit/496e045))
+* **f-date:** import flatpickr styles globally ([f7de9cf](https://github.com/pascaliske/form-elements/commit/f7de9cf))
+* **f-icon:** remove f-icon and use fa-icon directly ([b22e97a](https://github.com/pascaliske/form-elements/commit/b22e97a))
+* **f-input:** fix f-input stories ([066da20](https://github.com/pascaliske/form-elements/commit/066da20))
+* **f-text-area:** move styles into shared mixins ([c017264](https://github.com/pascaliske/form-elements/commit/c017264))
+* **package:** update core-js to version 3.0.0 ([7992c9c](https://github.com/pascaliske/form-elements/commit/7992c9c))
+* **storybook:** improve storybook style loading ([7c20cdc](https://github.com/pascaliske/form-elements/commit/7c20cdc))
+* **storybook:** set base font size for alle elements ([95c05e5](https://github.com/pascaliske/form-elements/commit/95c05e5))
+* **storybook:** sort stories alphabetically ([08544b1](https://github.com/pascaliske/form-elements/commit/08544b1))
+* **styles:** fix global form field styles ([2f5cf81](https://github.com/pascaliske/form-elements/commit/2f5cf81))
+* **styles:** remove legacy import ([4e1cd08](https://github.com/pascaliske/form-elements/commit/4e1cd08))
+* **styles:** update main styles ([b7a3dcd](https://github.com/pascaliske/form-elements/commit/b7a3dcd))
+* remove typings file and add typings directly to components ([e2d217f](https://github.com/pascaliske/form-elements/commit/e2d217f))
+
+
+### Code Refactoring
+
+* move dependencies to peer dependencies ([94aa1fa](https://github.com/pascaliske/form-elements/commit/94aa1fa))
+
+
+### Features
+
+* **f-button:** add stories for f-button component and optimize styles ([0a0511b](https://github.com/pascaliske/form-elements/commit/0a0511b))
+* **f-checkbox:** add stories for f-checkbox component and optimize styles ([df5f73f](https://github.com/pascaliske/form-elements/commit/df5f73f))
+* **f-column:** add stories for f-column component and optimize styles ([65f1598](https://github.com/pascaliske/form-elements/commit/65f1598))
+* **f-date:** add stories for f-date component and optimize styles ([d3afc1f](https://github.com/pascaliske/form-elements/commit/d3afc1f))
+* **f-email:** add stories for f-email component and optimize styles ([ddd7879](https://github.com/pascaliske/form-elements/commit/ddd7879))
+* **f-input:** add stories for f-input ([8d476fd](https://github.com/pascaliske/form-elements/commit/8d476fd))
+* **f-password:** add stories for f-password component and optimize styles ([97a5549](https://github.com/pascaliske/form-elements/commit/97a5549))
+* **f-phone:** add stories for f-phone component and optimize styles ([42a7ca2](https://github.com/pascaliske/form-elements/commit/42a7ca2))
+* **f-radiobutton:** add stories for f-radiobutton component and optimize styles ([5ec5a04](https://github.com/pascaliske/form-elements/commit/5ec5a04))
+* **f-row:** add stories for f-row component and optimize styles ([38c417b](https://github.com/pascaliske/form-elements/commit/38c417b))
+* **f-select:** add stories for f-select component and optimize styles ([bf763c1](https://github.com/pascaliske/form-elements/commit/bf763c1))
+* **f-text-area:** add stories for f-text-area component and optimize styles ([6d06f37](https://github.com/pascaliske/form-elements/commit/6d06f37))
+* **form-elements:** upgrade font awesome to v5 ([c0585a3](https://github.com/pascaliske/form-elements/commit/c0585a3))
+* **storybook:** add module dependencies for all stories ([bb36a04](https://github.com/pascaliske/form-elements/commit/bb36a04))
+* **storybook:** setup storybook for form-elements ([0fc8361](https://github.com/pascaliske/form-elements/commit/0fc8361))
+
+
+### BREAKING CHANGES
+
+* Remove direct dependency on flatpickr and fuzzysearch and move them to peerDependencies. Users should install them by themselves.
+
+
+
 ## [1.2.5](https://github.com/pascaliske/form-elements/compare/v1.2.4...v1.2.5) (2019-02-24)